refactor(frontend): unify file validation alerts
